Get the basic BTLE service up and running
This commit adds the basic framework for offering the U2F Bluetooth
service. Basically it adds the service and characteristics. It can
also reply correctly to reads of u2fControlPointLength and
u2fServiceRevisionBitfield. Everything else is unimplemented.

The good news is that u2ftest can recognize the phone with this code
installed as if it were a U2F token. Next step: fill in the stubs with
actual protocol responses.